+++ examples/feature.rs
@@ -216,6 +216,7 @@ fn single(name: &str) -> Result<()> {
         "image::push_texel_quater_no_bg" => {
             image::show_push_texel_quater_no_bg()
         }
+        "raw::request_ambiguous" => raw::show_request_ambiguous(),
         _ => {
             eprintacln!("{'r}error: {'_}unknown feature `{name}`.");
             Ok(())
@@ -396,6 +397,7 @@ When running with cargo, instead of `{'c}feature{'_}` you use
   {'c}image{'y}::{'w i}push_texel_quater{'_}
   {'c}image{'y}::{'w i}push_texel_half_no_bg{'_}
   {'c}image{'y}::{'w i}push_texel_quater_no_bg{'_}
+  {'c}raw{'y}::{'w i}request_ambiguous{'_}
     "
     );
     Ok(())